Tidy tests
