+++ fitsio/src/fitsfile.rs
@@ -274,7 +274,7 @@ impl FitsFile {
     }
 
     /// Return the list of HDU names
-    fn hdu_names(&mut self) -> Result<Vec<String>> {
+    pub(crate) fn hdu_names(&mut self) -> Result<Vec<String>> {
         let num_hdus = self.num_hdus()?;
         let mut result = Vec::with_capacity(num_hdus);
         for i in 0..num_hdus {
@@ -1454,61 +1454,6 @@ mod test {
     }
 
     #[test]
-    fn test_read_image_data() {
-        let mut f = FitsFile::open("../testdata/full_example.fits").unwrap();
-        let hdu = f.hdu(0).unwrap();
-        let first_row: Vec<i32> = hdu.read_section(&mut f, 0, 100).unwrap();
-        assert_eq!(first_row.len(), 100);
-        assert_eq!(first_row[0], 108);
-        assert_eq!(first_row[49], 176);
-
-        let second_row: Vec<i32> = hdu.read_section(&mut f, 100, 200).unwrap();
-        assert_eq!(second_row.len(), 100);
-        assert_eq!(second_row[0], 177);
-        assert_eq!(second_row[49], 168);
-    }
-
-    #[test]
-    fn test_read_whole_image() {
-        let mut f = FitsFile::open("../testdata/full_example.fits").unwrap();
-        let hdu = f.hdu(0).unwrap();
-        let image: Vec<i32> = hdu.read_image(&mut f).unwrap();
-        assert_eq!(image.len(), 10000);
-    }
-
-    #[test]
-    fn test_read_image_rows() {
-        let mut f = FitsFile::open("../testdata/full_example.fits").unwrap();
-        let hdu = f.hdu(0).unwrap();
-        let row: Vec<i32> = hdu.read_rows(&mut f, 0, 2).unwrap();
-        let ref_row: Vec<i32> = hdu.read_section(&mut f, 0, 200).unwrap();
-        assert_eq!(row, ref_row);
-    }
-
-    #[test]
-    fn test_read_image_row() {
-        let mut f = FitsFile::open("../testdata/full_example.fits").unwrap();
-        let hdu = f.hdu(0).unwrap();
-        let row: Vec<i32> = hdu.read_row(&mut f, 0).unwrap();
-        let ref_row: Vec<i32> = hdu.read_section(&mut f, 0, 100).unwrap();
-        assert_eq!(row, ref_row);
-    }
-
-    #[test]
-    fn test_read_image_slice() {
-        let mut f = FitsFile::open("../testdata/full_example.fits").unwrap();
-        let hdu = f.hdu(0).unwrap();
-
-        let xcoord = 5..7;
-        let ycoord = 2..3;
-
-        let chunk: Vec<i32> = hdu.read_region(&mut f, &vec![&ycoord, &xcoord]).unwrap();
-        assert_eq!(chunk.len(), 2);
-        assert_eq!(chunk[0], 168);
-        assert_eq!(chunk[chunk.len() - 1], 193);
-    }
-
-    #[test]
     fn test_read_image_region_from_table() {
         let mut f = FitsFile::open("../testdata/full_example.fits").unwrap();
         let hdu = f.hdu("TESTEXT").unwrap();
@@ -1529,164 +1474,6 @@ mod test {
         } else {
             panic!("Should have been an error");
         }
-    }
-
-    #[test]
-    fn test_write_image_section() {
-        with_temp_file(|filename| {
-            let data_to_write: Vec<i64> = (0..100).map(|v| v + 50).collect();
-
-            // Scope ensures file is closed properly
-            {
-                use fitsfile::ImageDescription;
-
-                let mut f = FitsFile::create(filename).open().unwrap();
-                let image_description = ImageDescription {
-                    data_type: ImageType::Long,
-                    dimensions: &[100, 20],
-                };
-                let hdu = f.create_image("foo".to_string(), &image_description)
-                    .unwrap();
-                hdu.write_section(&mut f, 0, 100, &data_to_write).unwrap();
-            }
-
-            let mut f = FitsFile::open(filename).unwrap();
-            let hdu = f.hdu("foo").unwrap();
-            let first_row: Vec<i64> = hdu.read_section(&mut f, 0, 100).unwrap();
-            assert_eq!(first_row, data_to_write);
-        });
-    }
-
-    #[test]
-    fn test_write_image_region() {
-        with_temp_file(|filename| {
-            // Scope ensures file is closed properly
-            {
-                use fitsfile::ImageDescription;
-
-                let mut f = FitsFile::create(filename).open().unwrap();
-                let image_description = ImageDescription {
-                    data_type: ImageType::Long,
-                    dimensions: &[100, 5],
-                };
-                let hdu = f.create_image("foo".to_string(), &image_description)
-                    .unwrap();
-
-                let data: Vec<i64> = (0..66).map(|v| v + 50).collect();
-                hdu.write_region(&mut f, &[&(0..10), &(0..5)], &data)
-                    .unwrap();
-            }
-
-            let mut f = FitsFile::open(filename).unwrap();
-            let hdu = f.hdu("foo").unwrap();
-            let chunk: Vec<i64> = hdu.read_region(&mut f, &[&(0..10), &(0..5)]).unwrap();
-            assert_eq!(chunk.len(), 10 * 5);
-            assert_eq!(chunk[0], 50);
-            assert_eq!(chunk[25], 75);
-        });
-    }
-
-    #[test]
-    fn test_write_image() {
-        with_temp_file(|filename| {
-            let data: Vec<i64> = (0..2000).collect();
-
-            // Scope ensures file is closed properly
-            {
-                use fitsfile::ImageDescription;
-
-                let mut f = FitsFile::create(filename).open().unwrap();
-                let image_description = ImageDescription {
-                    data_type: ImageType::Long,
-                    dimensions: &[100, 20],
-                };
-                let hdu = f.create_image("foo".to_string(), &image_description)
-                    .unwrap();
-
-                hdu.write_image(&mut f, &data).unwrap();
-            }
-
-            let mut f = FitsFile::open(filename).unwrap();
-            let hdu = f.hdu("foo").unwrap();
-            let chunk: Vec<i64> = hdu.read_image(&mut f).unwrap();
-            assert_eq!(chunk, data);
-        });
-    }
-
-    #[test]
-    fn test_resizing_images() {
-        with_temp_file(|filename| {
-            // Scope ensures file is closed properly
-            {
-                use fitsfile::ImageDescription;
-
-                let mut f = FitsFile::create(filename).open().unwrap();
-                let image_description = ImageDescription {
-                    data_type: ImageType::Long,
-                    dimensions: &[100, 20],
-                };
-                f.create_image("foo".to_string(), &image_description)
-                    .unwrap();
-            }
-
-            /* Now resize the image */
-            {
-                let mut f = FitsFile::edit(filename).unwrap();
-                let hdu = f.hdu("foo").unwrap();
-                hdu.resize(&mut f, &[1024, 1024]).unwrap();
-            }
-
-            /* Images are only resized when flushed to disk, so close the file and
-             * open it again */
-            {
-                let mut f = FitsFile::edit(filename).unwrap();
-                let hdu = f.hdu("foo").unwrap();
-                match hdu.info {
-                    HduInfo::ImageInfo { shape, .. } => {
-                        assert_eq!(shape, [1024, 1024]);
-                    }
-                    _ => panic!("Unexpected hdu type"),
-                }
-            }
-        });
-    }
-
-    #[test]
-    fn test_resize_3d() {
-        with_temp_file(|filename| {
-            // Scope ensures file is closed properly
-            {
-                use fitsfile::ImageDescription;
-
-                let mut f = FitsFile::create(filename).open().unwrap();
-                let image_description = ImageDescription {
-                    data_type: ImageType::Long,
-                    dimensions: &[100, 20],
-                };
-                f.create_image("foo".to_string(), &image_description)
-                    .unwrap();
-            }
-
-            /* Now resize the image */
-            {
-                let mut f = FitsFile::edit(filename).unwrap();
-                let hdu = f.hdu("foo").unwrap();
-                hdu.resize(&mut f, &[1024, 1024, 5]).unwrap();
-            }
-
-            /* Images are only resized when flushed to disk, so close the file and
-             * open it again */
-            {
-                let mut f = FitsFile::edit(filename).unwrap();
-                let hdu = f.hdu("foo").unwrap();
-                match hdu.info {
-                    HduInfo::ImageInfo { shape, .. } => {
-                        assert_eq!(shape, [1024, 1024, 5]);
-                    }
-                    _ => panic!("Unexpected hdu type"),
-                }
-            }
-        });
     }
 
     #[test]
@@ -1733,24 +1520,6 @@ mod test {
                 panic!("Should have thrown an error");
             }
         });
-    }
-
-    #[test]
-    fn test_multi_hdu_workflow() {
-        /* Check that hdu objects change the current HDU on every file access method */
-
-        let mut f = FitsFile::open("../testdata/full_example.fits").unwrap();
-        let primary_hdu = f.hdu(0).unwrap();
-        let column_hdu = f.hdu(1).unwrap();
-
-        let first_row: Vec<i32> = primary_hdu.read_section(&mut f, 0, 100).unwrap();
-        assert_eq!(first_row.len(), 100);
-        assert_eq!(first_row[0], 108);
-        assert_eq!(first_row[49], 176);
-
-        let intcol_data: Vec<i32> = column_hdu.read_col(&mut f, "intcol").unwrap();
-        assert_eq!(intcol_data[0], 18);
-        assert_eq!(intcol_data[49], 12);
     }
 
     #[test]
@@ -1809,145 +1578,4 @@ mod test {
         });
     }
 
-    #[test]
-    fn test_fetch_hdu_name() {
-        duplicate_test_file(|filename| {
-            let mut f = FitsFile::open(filename).unwrap();
-            let hdu = f.hdu("TESTEXT").unwrap();
-            assert_eq!(hdu.name(&mut f).unwrap(), "TESTEXT".to_string());
-        });
-    }
-
-    #[test]
-    fn test_inserting_columns() {
-        duplicate_test_file(|filename| {
-            let mut f = FitsFile::edit(filename).unwrap();
-            let hdu = f.hdu("TESTEXT").unwrap();
-
-            let coldesc = ColumnDescription::new("abcdefg")
-                .with_type(ColumnDataType::Int)
-                .create()
-                .unwrap();
-
-            let newhdu = hdu.insert_column(&mut f, 0, &coldesc).unwrap();
-
-            match newhdu.info {
-                HduInfo::TableInfo {
-                    column_descriptions,
-                    ..
-                } => {
-                    assert_eq!(column_descriptions[0].name, "abcdefg");
-                }
-                _ => panic!("ERROR"),
-            }
-        });
-    }
-
-    #[test]
-    fn test_appending_columns() {
-        duplicate_test_file(|filename| {
-            let mut f = FitsFile::edit(filename).unwrap();
-            let hdu = f.hdu("TESTEXT").unwrap();
-
-            let coldesc = ColumnDescription::new("abcdefg")
-                .with_type(ColumnDataType::Int)
-                .create()
-                .unwrap();
-
-            let newhdu = hdu.append_column(&mut f, &coldesc).unwrap();
-
-            match newhdu.info {
-                HduInfo::TableInfo {
-                    column_descriptions,
-                    ..
-                } => {
-                    assert_eq!(
-                        column_descriptions[column_descriptions.len() - 1].name,
-                        "abcdefg"
-                    );
-                }
-                _ => panic!("ERROR"),
-            }
-        });
-    }
-
-    #[test]
-    fn test_deleting_columns_by_name() {
-        duplicate_test_file(|filename| {
-            let mut f = FitsFile::edit(filename).unwrap();
-            let hdu = f.hdu("TESTEXT").unwrap();
-            let newhdu = hdu.delete_column(&mut f, "intcol").unwrap();
-
-            match newhdu.info {
-                HduInfo::TableInfo {
-                    column_descriptions,
-                    ..
-                } => for col in column_descriptions {
-                    assert!(col.name != "intcol");
-                },
-                _ => panic!("ERROR"),
-            }
-        });
-    }
-
-    #[test]
-    fn test_delete_hdu() {
-        duplicate_test_file(|filename| {
-            {
-                let mut f = FitsFile::edit(filename).unwrap();
-                let hdu = f.hdu("TESTEXT").unwrap();
-                hdu.delete(&mut f).unwrap();
-            }
-
-            let mut f = FitsFile::open(filename).unwrap();
-            let hdu_names = f.hdu_names().unwrap();
-            assert!(!hdu_names.contains(&"TESTEXT".to_string()));
-        });
-    }
-
-    #[test]
-    fn test_deleting_columns_by_number() {
-        duplicate_test_file(|filename| {
-            let mut f = FitsFile::edit(filename).unwrap();
-            let hdu = f.hdu("TESTEXT").unwrap();
-            let newhdu = hdu.delete_column(&mut f, 0).unwrap();
-
-            match newhdu.info {
-                HduInfo::TableInfo {
-                    column_descriptions,
-                    ..
-                } => for col in column_descriptions {
-                    assert!(col.name != "intcol");
-                },
-                _ => panic!("ERROR"),
-            }
-        });
-    }
-
-    #[test]
-    fn test_hdu_iterator() {
-        duplicate_test_file(|filename| {
-            let mut f = FitsFile::open(filename).unwrap();
-            let mut counter = 0;
-
-            for _ in f.iter() {
-                counter += 1;
-            }
-
-            assert_eq!(counter, 2);
-        });
-    }
-
-    #[test]
-    fn test_read_single_table_value() {
-        let filename = "../testdata/full_example.fits[TESTEXT]";
-        let mut f = FitsFile::open(filename).unwrap();
-        let tbl_hdu = f.hdu("TESTEXT").unwrap();
-
-        let result: i64 = tbl_hdu.read_cell_value(&mut f, "intcol", 4).unwrap();
-        assert_eq!(result, 16);
-
-        let result: String = tbl_hdu.read_cell_value(&mut f, "strcol", 4).unwrap();
-        assert_eq!(result, "value4".to_string());
-    }
 }
